Improve debug logging
This also includes one fix that I discovered while improving the
logging. Even if a git url was provided, release_gitter was lookig for a
local package declaration (Cargo.toml) to identify the version.

With this change, the url parsing and the local repo logic are split
allowing for more detailed logging as well as avoiding this potential
bug.